feat(console): migrate "Alerts" into new menu for API v2
