+++ gravitee-apim-console-webui/src/components/alerts/alert/triggers/trigger-application-quota.component.ts
@@ -21,7 +21,7 @@ const AlertTriggerApplicationQuotaComponent: ng.IComponentOptions = {
     alert: '<',
   },
   require: {
-    parent: '^alertComponent',
+    parent: '^alertComponentAjs',
   },
   template: require('./trigger-application-quota.html'),
   /* @ngInject */
