+++ gravitee-apim-console-webui/src/components/alerts/alert/triggers/trigger-metrics-aggregation.component.ts
@@ -22,7 +22,7 @@ const AlertTriggerMetricsAggregationComponent: ng.IComponentOptions = {
     alert: '<',
   },
   require: {
-    parent: '^alertComponent',
+    parent: '^alertComponentAjs',
   },
   template: require('./trigger-metrics-aggregation.html'),
   /* @ngInject */
