+++ gravitee-apim-console-webui/src/components/alerts/dashboard/alerts-dashboard.component.ts
@@ -124,8 +124,8 @@ class AlertsDashboardComponent implements ng.IComponentController {
       case 'API':
         return {
           scope: Scope.API,
-          alertCreationUiRef: 'management.apis.detail.alerts.alertnew',
-          uiRef: 'management.apis.detail.alerts.editalert({alertId: alert.id, tab: "history"})',
+          alertCreationUiRef: 'management.apis.ng.alerts.alertnew',
+          uiRef: 'management.apis.detail.ng.editalert({alertId: alert.id, tab: "history"})',
           permission: 'api-alert-r',
           hasPermission: this.UserService.currentUser?.userApiPermissions.includes('api-alert-r'),
         };
